niwot - client/requirements.txt
PySide6==6.7.2
requests==2.32.3
python-socketio[client]==5.11.3
websocket-client==1.8.0
requests-toolbelt==1.0.0
